+++ WebShop/src/main/java/com/webShop/WebShop/repository/TransactionRepository.java
@@ -12,5 +12,5 @@ public interface TransactionRepository extends JpaRepository<Transaction, Long> 
     @Query(value= "select t from Transaction t join fetch User u where u.id = ?1", nativeQuery = true)
     List<Transaction> findTransactionsForUser(long id);
     Transaction findByPaymentId(String paymentId);
-    Transaction findByMerchantOrderId(long merchantOrderId);
+    Transaction findByMerchantOrderId(int merchantOrderId);
 }
